+++ TractorMod/Framework/CustomSaveBuilding.cs
@@ -14,6 +14,9 @@ namespace TractorMod.Framework
         /// <summary>The tile location.</summary>
         public Vector2 Tile { get; }
 
+        /// <summary>The number of days until construction ends.</summary>
+        public int DaysOfConstructionLeft { get; }
+
 
         /*********
         ** Public methods
@@ -21,10 +24,12 @@ namespace TractorMod.Framework
         /// <summary>Construct an instance.</summary>
         /// <param name="tile">The building type.</param>
         /// <param name="type">The tile location.</param>
-        public CustomSaveBuilding(Vector2 tile, string type)
+        /// <param name="daysOfConstructionLeft">The number of days until construction ends.</param>
+        public CustomSaveBuilding(Vector2 tile, string type, int daysOfConstructionLeft)
         {
             this.Tile = tile;
             this.Type = type;
+            this.DaysOfConstructionLeft = daysOfConstructionLeft;
         }
     }
 }